feat: implement server for db
